Refactor to move DefaultLogger to middlewares package.
